requirements.txt
flask==3.0.0
psycopg2-binary==2.9.9
numpy==1.26.0
requests==2.31.0
matplotlib==3.8.2
geopy==2.4.1